Add files via upload
Added BinarySearchTree to Tree.py.
Uses a BinaryNode subclass to organize data.

I'm not particularly proud of the organization of this class. Because every sub-tree in a binary search tree is itself a binary search tree,
it makes more sense in hindsight to ditch the BinarySearchTree class -> BinaryNode sub-class in favor of a single-class. This would have the added benefit of making recursive methods easier to implement.